feat: move exhibition/enter,exit paths to guests/enter,exit
