+++ src/apis/guests/_id@string/exit/index.ts
@@ -1,5 +1,5 @@
 import { AuthToken } from "../../../@types";
-import { Guest } from "../../general/guest/@types";
+import { Guest } from "../../@types";
 
 export interface Methods {
   /**
